Charm
  - Moved CkReductionMsg-related definitions from main_enzo.cpp to
    charm_reductions.?pp and enzo_reductions.?pp

Cleaning

  - Addressed numerous compile-time warnings
  - Fixed some cases where std::numeric_limits<float>::min()
    should really be -max()

Testing

  - Enhancing test_Field.cpp with testing temporary fields
    (NOTE: tests fail)

Solver

  - Changed EnzoSolverCg enzo_block->is_leaf() to is_active_(block)
  - Changed Solver::neighbor_type_(), sync_type_(), and is_active_()
    to be public
  - Fixed logic in EnzoSolverMg0 from (!is_leaf() || level<max_level)
    to (!is_leaf() && level < max_level)
  - Changed checking precision of "density" field to 0 field
  - Cleaned EnzoSolverMg0::[prolong|restrict]_recv(), moving data copies
    from EnzoBlock callback to EnzoSolverMg0.
  - Bug Fix in EnzoSolverMg0--had hard-coded "B" and "C", though "B" is
    now set in ::apply() method
  - Added FieldMsg * to EnzoSolverMg0::[restrict|prolong]_recv()

Data
  - Bug fixes related to temporary fields
  - Updated test_Field tests for temporary fields

--HG--
branch : solver

+++ src/Cello/charm_MsgRefine.cpp
@@ -185,7 +185,7 @@ void * MsgRefine::pack (MsgRefine * msg)
   (*pi++) = msg->num_face_level_;
 
   // face_level_[]
-  for (size_t i=0; i<msg->num_face_level_; i++) {
+  for (int i=0; i<msg->num_face_level_; i++) {
     (*pi++) = msg->face_level_[i];
   }
 
@@ -272,7 +272,7 @@ MsgRefine * MsgRefine::unpack(void * buffer)
   // face_level_[]
   if (msg->num_face_level_ > 0) {
     msg->face_level_ = new int [msg->num_face_level_];
-    for (size_t i = 0; i<msg->num_face_level_; i++) {
+    for (int i = 0; i<msg->num_face_level_; i++) {
       msg->face_level_[i] = (*pi++);
     }
   } else {
